WMV v0.7.0.1 -64-bit Character fixes. -Added Documentation for Cataclysm ADT files.
WMV v0.8.0.0
-Added Animation Controls layout.
-Added preliminary Exporters dll project.
-Added StormLib 8.0.1.
-Added more Content-awareness to the main menu.
-Certain menu options will now be disabled if WoW is not loaded.
-Continued work on the Model headers.
-The File List is now openable & closable from the View->Controls menu.
-Rebuilt .pro files.

+++ trunk/src/database.cpp
@@ -254,7 +254,7 @@ CharRacesDB::Record CharRacesDB::getByName(wxString name)
 	throw NotFound();
 }
 
-CharRacesDB::Record CharRacesDB::getById(unsigned int id)
+CharRacesDB::Record CharRacesDB::getById(size_t id)
 {
 	for(Iterator i=begin(); i!=end(); ++i)
 	{
